ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
yfinance==0.2.18
requests==2.31.0
beautifulsoup4==4.12.2
textblob==0.17.1
nltk==3.8.1
python-dotenv==1.0.0
pydantic==2.5.0
schedule==1.2.0
ta==0.10.2
plotly==5.17.0
dash==2.14.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2 